Adding Error handlers and validation delete popups for all delete actions

+++ Application/DistributorManagment/Commands/DeleteDistributorUser/DeleteDistributorUserCommand.cs
@@ -1,4 +1,5 @@
-﻿using Domain.DistributorManagment.AggregatesModel.DistributorAggregate;
+﻿using Domain.Common.Exceptions;
+using Domain.DistributorManagment.AggregatesModel.DistributorAggregate;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -28,6 +29,7 @@ namespace Application.DistributorManagment.Commands.DeleteDistributorUser
             {
                 // get distributor by id
                 var distributorFromRepo = await _distributorRepository.FindByIdAsync(request.DistributorId);
+                if (distributorFromRepo == null) if (distributorFromRepo == null) throw new RestException(System.Net.HttpStatusCode.BadRequest, new { Distributor = $"Distributor with id {request.DistributorId} not found ", code = "distributor_notfound" });
 
                 // delete distributorUser to distributor
                 distributorFromRepo.DeleteDistributorUser(request.DistributorUserId);
