+++ Application/ProductCatalog/ProductAggregate/Commands/DeleteUnit/DeleteUnitCommand.cs
@@ -1,4 +1,6 @@
-﻿using Domain.ProductCatalog.AggregatesModel.ProductAggregate;
+﻿using Application.Common.Exceptions;
+using Domain.Common.Exceptions;
+using Domain.ProductCatalog.AggregatesModel.ProductAggregate;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -28,6 +30,7 @@ namespace Application.ProductCatalog.ProductAggregate.Commands.DeleteUnit
             {
                 // get product by id
                 var productFromRepo = await _productRepository.FindByIdAsync(request.ProductId);
+                if (productFromRepo == null) throw new RestException(System.Net.HttpStatusCode.BadRequest, new { Product = $"Product with id {request.ProductId} not found ", code = "product_notfound" });
 
                 // delete unit to product
                 productFromRepo.DeleteProductUnit(request.UnitId);
